Metric Class

An R6 class representing a metric for evaluating Polygenic Scores (PGS).

Public fields

name

Character. The name of the metric.

fn

Function. The function used to calculate the metric.

type

binary/continuous/both.

description

Longer name.

Methods

Public methods

Metric$new()

Metric$calculate()

Metric$clone()

Method new()

Create a new Metric object.

Usage

Metric$new(name, fn, type, description)

Arguments

name

Character. The name of the metric.

fn

Function. The function used to calculate the metric.

type

binary/continuous/both.

description

Longer name.

Method calculate()

Calculate the metric value.

Usage

Metric$calculate(
  m = NULL,
  m_lm = NULL,
  p = NULL,
  y = NULL,
  liab_prep = NULL,
  n = NULL,
  w = NULL
)

Arguments

m

glm model

m_lm

lm model (family='gaussian')

p

probabilities from glm

y

response

liab_prep

data input to r2_liability

n

number of observations

w

weights

Returns

The calculated metric value.

Method clone()

The objects of this class are cloneable with this method.

Usage

Metric$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.